requirements.txt
annotated-types==0.6.0
anyio==4.3.0
camelot-py==0.11.0
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
decorator==5.1.1
distro==1.9.0
Elixir==0.7.1
et-xmlfile==1.1.0
exceptiongroup==1.2.1
fastapi==0.110.2
h11==0.14.0
idna==3.7
Jinja2==3.1.3
JPype1==1.5.0
MarkupSafe==2.1.5
numpy==1.26.4
opencv-python==4.9.0.80
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
pbr==6.0.0
pdfminer.six==20231228
psycopg2-binary
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
PyMySQL==1.1.0
pypdf==4.2.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytz==2024.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
sqlalchemy-migrate==0.11.0
sqlparse==0.5.0
starlette==0.37.2
tabula-py==2.9.0
tabulate==0.9.0
Tempita==0.5.2
typing_extensions==4.11.0
tzdata==2024.1
uvicorn==0.29.0
xlrd==0.7.1
xlwt==0.7.2
